Add the S3 bucket url to the env in the manifest
The federalist-proxy fields a large amount of traffic and as such needs to be deployed with autopilot. For this to work, environment variables need to be specified in either the manifest or a user provided service.

This commit adds the S3 bucket url to the manifest so that it doesn't need to be re-added every time we redeploy the proxy