Setup Docker so the proxy can run locally
This commit adds a Dockerfile and a Docker Compose configuration which can be used to run the app locally. The app will run as if it were running in the staging environment

As part of the above change, the app also adds a staging manifest so the app's staging configuration can be described.