Swap proxy redirects for /preview and /demo
This commit changes the proxy so that requests with the prefix
`/demo/<org>/<repo>` trim the `/demo/<org>/<repo>` element from
redirects. For requests with the prefix `/prefix/<org>/<repo>/<branch>`, it
quits trimming the prefix.

This is necessary now that demo branches are published in `/demo`
instead of `/preview`.