Remove proxy_redirect rules
After https://github.com/18F/federalist-garden-build/pull/47 the build container started creating redirect objects for the site. A result of this is that we no longer need to strip the baseurl out of demo and site paths since the BASEURL will handle that when the build is run.

This commit removes the proxy_redirect rules that strip the baseurl out of demo and site paths.